backend/requirements.txt
openai
python-dotenv
fastapi
uvicorn
tavily-python
supabase